+++ tests/test_recursion.py
@@ -322,13 +322,14 @@ def test_recursion_with_bad_item():
             "valid_stac": False,
             "error_type": "JSONSchemaValidationError",
             "error_message": "'id' is a required property",
+            "recommendation": "For more accurate error information, rerun with --verbose.",
         },
     ]
 
 
-def test_recursion_with_bad_item_verbose():
+def test_recursion_with_bad_item_trace_recursion():
     stac_file = "tests/test_data/v100/catalog-with-bad-item.json"
-    stac = stac_validator.StacValidate(stac_file, recursive=True, verbose=True)
+    stac = stac_validator.StacValidate(stac_file, recursive=True, trace_recursion=True)
     stac.run()
     assert not stac.valid
     assert len(stac.message) == 2
@@ -352,6 +353,7 @@ def test_recursion_with_bad_item_verbose():
             "valid_stac": False,
             "error_type": "JSONSchemaValidationError",
             "error_message": "'id' is a required property",
+            "recommendation": "For more accurate error information, rerun with --verbose.",
         },
     ]
 
@@ -377,6 +379,7 @@ def test_recursion_with_bad_child_collection():
             "validation_method": "recursive",
             "error_type": "JSONSchemaValidationError",
             "error_message": "'id' is a required property",
+            "recommendation": "For more accurate error information, rerun with --verbose.",
         }
     ]
 
@@ -399,5 +402,6 @@ def test_recursion_with_missing_collection_link():
             "validation_method": "recursive",
             "error_type": "JSONSchemaValidationError",
             "error_message": "'simple-collection' should not be valid under {}. Error is in collection ",
+            "recommendation": "For more accurate error information, rerun with --verbose.",
         },
     ]